requirements.txt
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==2.1.7
django-crispy-forms==1.7.2
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
httpie==2.2.0
idna==2.10
mysqlclient==2.0.1
psycopg2-binary==2.8.6
Pygments==2.7.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
